pyproject.toml
[tool.pdm]
[tool.pdm.dev-dependencies]
dev = [
    "black>=23.7.0",
]

[project]
name = ""
version = ""
description = ""
authors = [
    {name = "Gabriela Saraiva Soares", email = "gabrielasaraiva.04@gmail.com"},
]
dependencies = [
    "django>=4.2.4",
    "djangorestframework>=3.14.0",
    "django-cors-headers>=4.2.0",
    "djangorestframework-simplejwt>=5.2.2",
]
requires-python = ">=3.10"
license = {text = "MIT"}
